requirements.txt
Pillow==8.1.2
PyMuPDF==1.18.9
